Move grain stream subscription attributes to Abstractions
